Fix Chess Pieces Name String To Not Use String Format
Because the chess pieces are using a string format along with the name of the
chess pieces in order to get the name of the image. This is causing issues with
checking the name of the chess piece as well as it does not look to good when
looking at the code.