Add King Cell Indenture in ChessBoard Class
Add in a king cell indenture to ChessBoard class because the king chess
piece is the most important piece and needs to be monitor as the player
may not be able to move if the King is in danger and in order to tell if the
game is over when the king dies.

Also change the initial state in aSimpleChessGame class in to add the
king cell where it should be on the board in order to make the chess board
in memory mint the actually board that the user is using.