Add All Chess Pieces As Chess Cell State
Add all of the chess pieces to be a state in a chess cell because it is more
apparent that the chess game in memory needs to know not only that it is
a chess piece but also what type of chess piece in order to display correct
one to the user. Also, update all of the unit test in order for them to use the
new chess cell states.