Rename addChessBoard() In SceneBoard file
Rename addChessBoard() to addChessBoardToDisplayForSize() in order to
be clear of what this function is actually doing. Also, add in some
comments to better id where in the file you are in because it will get
hard to know where are in the file as it gets bigger.  Furthermore,
came to the realization that all functions that have to with displaying
items on the screen should be in the SceneBoard file.