Move Memory Board From SceneBoard To View File
In order to make the game more MVC friendly for better reading of the
code and data management. Move the memory board from SceneBoard file
which is the view to the controller which will do that actually
manipulating of the memory board.

Also fix an issue with asimplechessgameTests using the old name of
setInitialGameState()