Change Chess Board Class To Use Chess Piece Class
Change Chess Board class to use Chess Piece class instead of NSInteger
because this will better represent what is on the screen.

Also, the class structure now does like this Chess Piece can only be
access throw Chess Board since a chess piece is on the chess board and
it controls the array that represents it in memory. ASimpleChessGame
class controls the Chess Board since it does the game data such as
who’s turn it is and the number of chess pieces not in play.