Remove Init & initChessBoardState From Chess Board
Remove functions Init() from Chess Board class because the init
function is not need since clearBoard() needs to return the items that
will be display and a init will not be able to do this. Also remove
initChessBoardState() because it is not really needed since the board
needs to be clear first. It is best to have clearBoard() do both the
clearing and the replaying of items to the board then having it split
up.