Add findAllCellState() To Chess Board Class
Add findAllCellState() to chess board class because this will add in
all of the cells for this board in order to be displayed.

Also, move chess board class file into a folder called SuperClass since
it is a superclass for aSimpleChessGame class. This is order to be more
organized of what each class does what.