Add Function To Get Allowed Moves In Chess Board
Add function to get all allowed moves for a chess piece in Chess Board
class because if the user selects a chess piece on the board they need
to know what moves that piece can do in order to play the game. The
actually exploring of where the pieces can go is done by Chess Move
class in order to have clear looking code. Also, only the rook moves is
done for now.