Change Array With Allowed Moves For Chess Piece
Change the data in the array with all of the moves for a chess piece,
in Chess Board, from being an another array of the location to a chess
piece object, for that cell. Because it will have move information
about the cell then just the location. Such as what type of chess piece
is on that cell, as this may be important later on.