Fix Moves That Subtract To Stop At Max Moves
Because the move functions for both Bottom and Left subtract from the
origin spot. The calculations to find how many moves needs to be from
origin spot tack away from the number of moves. In order to make sure
that the function stops at the max number of moves otherwise it will be
a negative number.