Add findMovesForChessPiece() To Chess Board
Add function findMovesForChessPiece() to Chess Board class because the
same code is being used in multiple places when getting the movements
this will simplify the search for the moves a bit.

Also, add in private enum in order to make it easier to find which side
to look at instead of memorizing the numbers. Remove
checkThereIsADestinationRow() since it is no longer being access in
multiple spots in the code.