Change Clear Board In Chess Board Class
Change Clear Board function in Chess Board class to be dynamic and
allow either White or Black to be at the bottom of the chess board
meaning that it is what the user has selected. This is because later on
the user will have the option of choose a which colour to selected.