Fix Travis CI Config To Use Debug
Fix Travis CI config to use debug inserted of release because
asimplechessgameTests which is where all of the unit testing are. Can
only use test build and since test build will using debug config. This
causes an error when Travis CI tries to run it as a release build
instead of a Test build.