Add Get Possible Moves Function For Bishop
Add in a function to get all of the possible moves for the bishop
because the user needs to see what move are available when the touch
the bishop on the screen.

Also, add in the rest of the cases in findMovesForChessPiece() that
will check all four corners of the chess piece for possible moves. In
order to get all possible moves for the bishop, queen, king and allow
the pawn to kill its opponents.