Fix Only Checking One Side For Castling
When doing a check if the King is able to do a castling on either side
the app would only check on side and not the other. Leaving the user
unable to do a castling on the kingside.

Also, Xcode update Info.plist file it is unknown why it did this.