Add Get Possible Moves For Knight
Add in getting of all possible moves for the knight because it is need
to give user options of what move they want to do. Also, it is
different from the other chess pieces because the knight is the only
chess piece to do a L move in order to move.