Fix Check For All Dangler To Use Correct Check
Fix if statements for all the dangler functions to use the correct
check, for example, if a row is being subtracted and columns is being
added then the row needs to make sure that it is above zero and column
is not bigger then the number of columns on the board. Because it is
pointless to check on rows or columns that not going the same direction.