Add Check If Cells Are Empty Front Of Chess Piece
Add in a function in chess move class to check if the cells in front of
a chess piece are empty. Because pawns only can go one way, however
they cannot kill what is in front of them. This check will make sure
there is only empty spaces up until a chess piece is found or the
number of moves that are allowed are used up.