Move checkForFriendOrFoe() From Private To Public
Move checkForFriendOrFoeOnRowOrColumn() in Chess Moves class from
private to public because when checking if a pawn has enemies dangle in
front of them they only need to change the columns as the rows will
only change once and will not change again, which checking the columns.
Because of this checkForFriendOrFoeOnRowOrColumn() is the best choose
since it does not change the row but only the column. Thus making it
easier to check both the dangler sides in front the pawn.  Furthermore,
will be keeping checkForFriendOrFoe() dangler version public, for now,
because it maybe useful in the future.