Move Who's Turn It Is To Chess Board Class
Move who's turn it is to Chess Board class from asimplechessgame class
because the players turn is a basic chess function that and thus it
should actually be in the Chess Board class since this is the class for
all basic functions of a chess game. Also add in a property in order to
know if any of the players are in check for their next turn.