Merge Duplicated Set Chess Piece In Chess Board
Merge duplicated set chess piece functions together in chess board
class because they were doing the same thing the only different was
that one did not return anything. Also, now that setCellState is the
public, this can be used when promoting pawns if they reach the other
side of the board.